Minor Edits
